requirements.txt
discord.py
python-docx
docxtpl
docx2pdf